More descriptive README and moved files that are confusing users
This change adds more data to the README around the structure of the
project as well as moves the old logo and weblogo files into an asset
archive.

The goal of this PR is to make it a bit more clear on what is what for
new developers as well as to give ourselves a good sense of what goes
where.

Tried my hand on creating a quick instructions for new kraefta pages.